re_viewer_context
Module cache

Viewer caches

Caches are registered lazily upon first use, see Caches::entry. The concrete caches exposed here are always available for all viewer crates.

Re-exports

pub use caches::Cache;

pub use caches::Caches;

pub use image_decode_cache::ImageDecodeCache;

pub use image_stats_cache::ImageStatsCache;

pub use tensor_stats_cache::TensorStatsCache;

pub use video_cache::VideoCache;

Modules

caches 🔒

image_decode_cache 🔒

image_stats_cache 🔒

tensor_stats_cache 🔒

video_cache 🔒