re_viewer_context::cache::caches
Trait Cache

pub trait Cache:
    Any
    + Send
    + Sync {
    // Required methods
    fn purge_memory(&mut self);
    fn as_any_mut(&mut self) -> &mut dyn Any;

    // Provided methods
    fn begin_frame(&mut self, _renderer_active_frame_idx: u64) { ... }
    fn on_store_events(&mut self, events: &[ChunkStoreEvent]) { ... }
}

A cache for memoizing things in order to speed up immediate mode UI & other immediate mode style things.

See also egus’s cache system, in [egui::cache] (https://docs.rs/egui/latest/egui/cache/index.html).

Required Methods

fn purge_memory(&mut self)

Attempt to free up memory.

fn as_any_mut(&mut self) -> &mut dyn Any

Converts itself to a mutable reference of Any, which enables mutable downcasting to concrete types.

Provided Methods

fn begin_frame(&mut self, _renderer_active_frame_idx: u64)

Called once per frame to potentially flush the cache.

_renderer_active_frame_idx: The global frame index as reported by re_renderer::RenderContext::active_frame_idx.

fn on_store_events(&mut self, events: &[ChunkStoreEvent])

React to the chunk store’s changelog, if needed.

Useful to e.g. invalidate unreachable data.