re_viewer_context::cache
Module video_stream_cache

Structs

PlayableVideoStream

Video stream from the store, ready for playback.

SampleBuffer 🔒

A buffer of multiple video sample data from the datastore.

VideoStreamCache

Caches metadata and active players for video streams.

VideoStreamCacheEntry 🔒

Entry in the video stream cache.

VideoStreamKey 🔒

Identifies a video stream.

Enums

VideoStreamProcessingError

Functions

adjust_gops_for_removed_samples_back 🔒

Adjust GOPs for removed samples at the back of the sample list.

adjust_gops_for_removed_samples_front 🔒

Adjust GOPs for removed samples at the front of the sample list.

load_video_data_from_chunks 🔒

read_samples_from_chunk 🔒

Reads all video samples from a chunk into an existing video description.

timescale_for_timeline 🔒

Type Aliases

SharablePlayableVideoStream