Module re_viewer_context::component_fallbacks

Enums

ComponentFallbackError

Error type for a fallback request.

ComponentFallbackProviderResult

Result for a fallback request to a provider.

Traits

ComponentFallbackProvider

Provides fallback values for components, implemented typically by crate::SpaceViewClass and crate::VisualizerSystem.

TypedComponentFallbackProvider

Provides a fallback value for a given component with known type.