Type Alias re_viewer_context::component_ui_registry::UntypedComponentEditOrViewCallback

type UntypedComponentEditOrViewCallback = Box<dyn Fn(&ViewerContext<'_>, &mut Ui, &dyn Array, EditOrView) -> Option<Box<dyn Array>> + Send + Sync>;

Callback for viewing, and maybe editing, a component via UI.

Draws a UI showing the current value and allows the user to edit it. If any edit was made, should return Some with the updated value. If no edit was made, should return None.

Aliased Type

struct UntypedComponentEditOrViewCallback(Unique<dyn Fn(&ViewerContext<'_>, &mut Ui, &(dyn Array + 'static), EditOrView) -> Option<Box<dyn Array>> + Sync + Send>, Global);

Fields
0: Unique<dyn Fn(&ViewerContext<'_>, &mut Ui, &(dyn Array + 'static), EditOrView) -> Option<Box<dyn Array>> + Sync + Send>1: Global