re_viewer_context
Enum ComponentFallbackProviderResult

pub enum ComponentFallbackProviderResult {
    Value(ArrayRef),
    ComponentNotHandled,
    SerializationError(SerializationError),
}

Result for a fallback request to a provider.

Variants

Value(ArrayRef)

A fallback value was successfully provided.

ComponentNotHandled

The fallback provider is not able to handle the given component.

This is not treated as an error and should be handled by looking up a placeholder value.

SerializationError(SerializationError)

Arrow serialization failed.

Unlike ComponentFallbackProviderResult::ComponentNotHandled, this is treated as an unexpected error.