Module re_viewer_context::global_context

Ultimately the goal is to have hierarchy of contexts from general to increasingly specific. Each part of the viewer should only have access a narrow context. Unfortunately, right now everything is still very intertwined with ViewerContext so we can’t pull the GlobalContext out of this crate yet.

Re-exports

pub use self::app_options::AppOptions;

pub use self::command_sender::command_channel;

pub use self::command_sender::CommandReceiver;

pub use self::command_sender::CommandSender;

pub use self::command_sender::SystemCommand;

pub use self::command_sender::SystemCommandSender;

pub use self::component_ui_registry::ComponentUiRegistry;

pub use self::component_ui_registry::ComponentUiTypes;

pub use self::item::Item;

Modules

app_options 🔒

command_sender 🔒

component_ui_registry 🔒

item 🔒

Structs

GlobalContext

Application context that is shared across all parts of the viewer.

Enums

DisplayMode

Which display mode are we currently in?