re_viewer_context::global_context::component_ui_registry
Type Alias UntypedComponentEditOrViewCallback

type UntypedComponentEditOrViewCallback = Box<dyn Fn(&ViewerContext<'_>, &mut Ui, &dyn Array, EditOrView) -> Option<ArrayRef> + Send + Sync>;

Callback for viewing, and maybe editing, a component via UI.

Draws a UI showing the current value and allows the user to edit it. If any edit was made, should return Some with the updated value. If no edit was made, should return None.

Aliased Type

struct UntypedComponentEditOrViewCallback(/* private fields */);