Module re_viewer_context::gpu_bridge::tensor_to_gpu

Upload tensors to re_renderer.

Enums

TextureKeyUsage ğŸ”’

Functions

cast_slice_to_cow ğŸ”’

class_id_tensor_to_gpu ğŸ”’

color_tensor_to_gpu ğŸ”’

depth_tensor_range ğŸ”’

depth_tensor_to_gpu ğŸ”’

general_texture_creation_desc_from_tensor ğŸ”’

Uploads the tensor to a texture in a format that closely resembled the input. Uses no Unorm/Snorm formats.

generate_texture_key ğŸ”’

Returns a texture key for a given row id & usage.

narrow_f64_to_f32s ğŸ”’

narrow_i64_to_f32s ğŸ”’

narrow_u64_to_f32s ğŸ”’

pad_and_cast ğŸ”’

pad_and_narrow_and_cast ğŸ”’

tensor_to_gpu

Set up tensor for rendering on the GPU.

texture_height_width_channels