Trait re_viewer_context::space_view::dyn_space_view_class::DynSpaceViewClass

pub trait DynSpaceViewClass {
    // Required methods
    fn name(&self) -> SpaceViewClassName;
    fn icon(&self) -> &'static Icon;
    fn help_text(&self, re_ui: &ReUi) -> WidgetText;
    fn on_register(
        &self,
        system_registry: &mut SpaceViewSystemRegistry
    ) -> Result<(), SpaceViewClassRegistryError>;
    fn new_state(&self) -> Box<dyn SpaceViewState>;
    fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>;
    fn preferred_tile_aspect_ratio(
        &self,
        state: &dyn SpaceViewState
    ) -> Option<f32>;
    fn layout_priority(&self) -> SpaceViewClassLayoutPriority;
    fn auto_spawn_heuristic(
        &self,
        _ctx: &ViewerContext<'_>,
        space_origin: &EntityPath,
        ent_paths: &PerSystemEntities
    ) -> AutoSpawnHeuristic;
    fn on_frame_start(
        &self,
        ctx: &mut ViewerContext<'_>,
        state: &mut dyn SpaceViewState,
        ent_paths: &PerSystemEntities,
        entity_properties: &mut EntityPropertyMap
    );
    fn selection_ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut dyn SpaceViewState,
        space_origin: &EntityPath,
        space_view_id: SpaceViewId
    );
    fn ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut dyn SpaceViewState,
        systems: &SpaceViewSystemRegistry,
        query: &ViewQuery<'_>
    );
}

Defines a class of space view without any concrete types making it suitable for storage and interfacing.

Implemented by crate::SpaceViewClass.

Each Space View in the viewer’s viewport has a single class assigned immutable at its creation time. The class defines all aspects of its behavior. It determines which entities are queried, how they are rendered, and how the user can interact with them.

Required Methods

fn name(&self) -> SpaceViewClassName

Name of this space view class.

Used for both ui display and identification. Must be unique within a viewer session.

TODO(#2336): Display name and identifier should be separate.

fn icon(&self) -> &'static Icon

Icon used to identify this space view class.

fn help_text(&self, re_ui: &ReUi) -> WidgetText

Help text describing how to interact with this space view in the ui.

fn on_register( &self, system_registry: &mut SpaceViewSystemRegistry ) -> Result<(), SpaceViewClassRegistryError>

Called once upon registration of the class

This can be used to register all built-in crate::ViewContextSystem and crate::ViewPartSystem.

fn new_state(&self) -> Box<dyn SpaceViewState>

Called once for every new space view instance of this class.

The state is not persisted across viewer sessions, only shared frame-to-frame.

fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>

Optional archetype of the Space View’s blueprint properties.

Blueprint components that only apply to the space view itself, not to the entities it displays.

fn preferred_tile_aspect_ratio(&self, state: &dyn SpaceViewState) -> Option<f32>

Preferred aspect ratio for the ui tiles of this space view.

fn layout_priority(&self) -> SpaceViewClassLayoutPriority

Controls how likely this space view will get a large tile in the ui.

fn auto_spawn_heuristic( &self, _ctx: &ViewerContext<'_>, space_origin: &EntityPath, ent_paths: &PerSystemEntities ) -> AutoSpawnHeuristic

Heuristic used to determine which space view is the best fit for a set of paths.

For each path in ent_paths, at least one of the registered crate::ViewPartSystem for this class returned true when calling crate::ViewPartSystem::heuristic_filter.

fn on_frame_start( &self, ctx: &mut ViewerContext<'_>, state: &mut dyn SpaceViewState, ent_paths: &PerSystemEntities, entity_properties: &mut EntityPropertyMap )

Executed for all active space views on frame start (before any ui is drawn), can be use for heuristic & state updates before populating the scene.

Is only allowed to access archetypes defined by Self::blueprint_archetype Passed entity properties are individual properties without propagated values.

fn selection_ui( &self, ctx: &mut ViewerContext<'_>, ui: &mut Ui, state: &mut dyn SpaceViewState, space_origin: &EntityPath, space_view_id: SpaceViewId )

Ui shown when the user selects a space view of this class.

TODO(andreas): Should this be instead implemented via a registered data_ui of all blueprint relevant types?

fn ui( &self, ctx: &mut ViewerContext<'_>, ui: &mut Ui, state: &mut dyn SpaceViewState, systems: &SpaceViewSystemRegistry, query: &ViewQuery<'_> )

Draws the ui for this space view type and handles ui events.

The state passed in was previously created by Self::new_state and is kept frame-to-frame.