Trait re_viewer_context::space_view::visualizer_system::VisualizerSystem

pub trait VisualizerSystem: Send + Sync + 'static {
    // Required methods
    fn visualizer_query_info(&self) -> VisualizerQueryInfo;
    fn execute(
        &mut self,
        ctx: &ViewerContext<'_>,
        query: &ViewQuery<'_>,
        view_ctx: &ViewContextCollection
    ) -> Result<Vec<QueueableDrawData>, SpaceViewSystemExecutionError>;
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn filter_visualizable_entities(
        &self,
        entities: ApplicableEntities,
        _context: &dyn VisualizableFilterContext
    ) -> VisualizableEntities { ... }
    fn applicability_filter(
        &self
    ) -> Option<Box<dyn VisualizerAdditionalApplicabilityFilter>> { ... }
    fn data(&self) -> Option<&dyn Any> { ... }
    fn initial_override_value(
        &self,
        _ctx: &ViewerContext<'_>,
        _query: &LatestAtQuery,
        _store: &DataStore,
        _entity_path: &EntityPath,
        _component: &ComponentName
    ) -> Option<DataCell> { ... }
}

Element of a scene derived from a single archetype query.

Is populated after scene contexts and has access to them.

Required Methods

fn visualizer_query_info(&self) -> VisualizerQueryInfo

Information about which components are queried by the visualizer.

fn execute( &mut self, ctx: &ViewerContext<'_>, query: &ViewQuery<'_>, view_ctx: &ViewContextCollection ) -> Result<Vec<QueueableDrawData>, SpaceViewSystemExecutionError>

Queries the data store and performs data conversions to make it ready for display.

Mustn’t query any data outside of the archetype.

fn as_any(&self) -> &dyn Any

Provided Methods

fn filter_visualizable_entities( &self, entities: ApplicableEntities, _context: &dyn VisualizableFilterContext ) -> VisualizableEntities

Filters a set of applicable entities (entities that have all required components), into to a set of visualizable entities.

The context passed in here is generated by crate::SpaceViewClass::visualizable_filter_context.

fn applicability_filter( &self ) -> Option<Box<dyn VisualizerAdditionalApplicabilityFilter>>

Additional filter for applicability.

If none is specified, applicability is solely determined by required components.

fn data(&self) -> Option<&dyn Any>

Optionally retrieves a data store reference from the scene element.

This is useful for retrieving data that is common to several visualizers of a crate::SpaceViewClass. For example, if most visualizers produce ui elements, a concrete crate::SpaceViewClass can pick those up in its crate::SpaceViewClass::ui method by iterating over all visualizers.

fn initial_override_value( &self, _ctx: &ViewerContext<'_>, _query: &LatestAtQuery, _store: &DataStore, _entity_path: &EntityPath, _component: &ComponentName ) -> Option<DataCell>

Returns an initial value to use when creating an override for a component for this visualizer. This is used as a fallback if the component doesn’t already have data.