re_viewer_context::store_hub
Struct StoreHub

pub struct StoreHub {
    persistence: BlueprintPersistence,
    active_entry: Option<StoreHubEntry>,
    active_application_id: Option<ApplicationId>,
    default_blueprint_by_app_id: HashMap<ApplicationId, StoreId>,
    active_blueprint_by_app_id: HashMap<ApplicationId, StoreId>,
    store_bundle: StoreBundle,
    table_stores: HashMap<TableId, TableStore>,
    should_enable_heuristics_by_app_id: HashSet<ApplicationId>,
    caches_per_recording: HashMap<StoreId, Caches>,
    blueprint_last_save: HashMap<StoreId, ChunkStoreGeneration>,
    blueprint_last_gc: HashMap<StoreId, ChunkStoreGeneration>,
}

Interface for accessing all blueprints and recordings

The StoreHub provides access to the EntityDb instances that are used to store both blueprints and recordings.

Internally, the StoreHub tracks which ApplicationId and recording id (StoreId) are currently active in the viewer. These can be configured using StoreHub::set_active_recording_id and StoreHub::set_active_app respectively.

Blueprints

For each ApplicationId, the StoreHub also keeps track of two blueprints:

The active blueprint

The default blueprint

Either on of these can be None.

The active blueprint is what the user would see and edit, if they were to select that app id. If there is no active blueprint, the default will be cloned and made active.

The default blueprint is usually the blueprint set by the SDK. This lets users reset the active blueprint to the one sent by the SDK.

Fields
persistence: BlueprintPersistence

How we load and save blueprints.

active_entry: Option<StoreHubEntry>active_application_id: Option<ApplicationId>default_blueprint_by_app_id: HashMap<ApplicationId, StoreId>active_blueprint_by_app_id: HashMap<ApplicationId, StoreId>store_bundle: StoreBundletable_stores: HashMap<TableId, TableStore>should_enable_heuristics_by_app_id: HashSet<ApplicationId>

These applications should enable the heuristics early next frame.

caches_per_recording: HashMap<StoreId, Caches>

Things that need caching.

blueprint_last_save: HashMap<StoreId, ChunkStoreGeneration>

The ChunkStoreGeneration from when the EntityDb was last saved

blueprint_last_gc: HashMap<StoreId, ChunkStoreGeneration>

The ChunkStoreGeneration from when the EntityDb was last garbage collected

Implementations

impl StoreHub

pub fn welcome_screen_app_id() -> ApplicationId

App ID used as a marker to display the welcome screen.

fn welcome_screen_blueprint_id() -> StoreId

Blueprint ID used for the default welcome screen blueprint

pub fn test_hub() -> Self

Used only for tests

pub fn new( persistence: BlueprintPersistence, setup_welcome_screen_blueprint: &dyn Fn(&mut EntityDb), ) -> Self

Create a new StoreHub.

The StoreHub will contain a single empty blueprint associated with the app ID returned by [StoreHub::welcome_screen_app_id]. It should be used as a marker to display the welcome screen.

pub fn store_bundle(&self) -> &StoreBundle

All the loaded recordings and blueprints.

pub fn read_context(&mut self) -> (StorageContext<'_>, Option<StoreContext<'_>>)

Get a read-only StorageContext and optionally a StoreContext (if available) from the StoreHub.

All of the returned references to blueprints and recordings will have a matching ApplicationId.

pub fn entity_db_mut(&mut self, store_id: &StoreId) -> &mut EntityDb

Mutable access to a EntityDb by id

pub fn insert_entity_db(&mut self, entity_db: EntityDb)

Insert a new recording or blueprint into the StoreHub.

Note that the recording is not automatically made active. Use StoreHub::set_active_recording_id if needed.

pub fn insert_table_store( &mut self, id: TableId, store: TableStore, ) -> Option<TableStore>

Inserts a new table into the store (potentially overwriting an existing entry).

fn remove_store(&mut self, store_id: &StoreId)

pub fn remove(&mut self, entry: &StoreHubEntry)

pub fn retain_recordings( &mut self, should_retain: impl FnMut(&EntityDb) -> bool, )

pub fn clear_entries(&mut self)

Remove all open recordings and applications, and go to the welcome page.

pub fn set_active_app(&mut self, app_id: ApplicationId)

Change the active ApplicationId.

Will ignore this request if the application id has no matching recording, unless no app id has been set yet at all so far.

pub fn close_app(&mut self, app_id: &ApplicationId)

Close this application and all its recordings.

pub fn active_app(&self) -> Option<&ApplicationId>

pub fn active_recording_id(&self) -> Option<&StoreId>

The recording id for the active recording.

pub fn active_recording(&self) -> Option<&EntityDb>

Directly access the EntityDb for the active recording.

pub fn active_entry(&self) -> Option<&StoreHubEntry>

Currently active entry if any.

pub fn active_caches(&self) -> Option<&Caches>

Directly access the Caches for the active recording.

This returns None only if there is no active recording: the cache itself is always present if there’s an active recording.

pub fn set_active_recording_id(&mut self, recording_id: StoreId)

Change the active/visible recording id.

This will also change the application-id to match the newly active recording.

pub fn set_activate_recording(&mut self, store_id: StoreId)

Activate a recording by its StoreId.

pub fn default_blueprint_id_for_app( &self, app_id: &ApplicationId, ) -> Option<&StoreId>

pub fn default_blueprint_for_app( &self, app_id: &ApplicationId, ) -> Option<&EntityDb>

pub fn set_default_blueprint_for_app( &mut self, app_id: &ApplicationId, blueprint_id: &StoreId, ) -> Result<()>

Change which blueprint is the default for a given ApplicationId

pub fn active_blueprint_id(&self) -> Option<&StoreId>

What is the active blueprint for the active application?

pub fn active_blueprint_id_for_app( &self, app_id: &ApplicationId, ) -> Option<&StoreId>

pub fn active_blueprint_for_app( &self, app_id: &ApplicationId, ) -> Option<&EntityDb>

pub fn set_cloned_blueprint_active_for_app( &mut self, app_id: &ApplicationId, blueprint_id: &StoreId, ) -> Result<()>

Make blueprint active for a given ApplicationId

We never activate a blueprint directly. Instead, we clone it and activate the clone.

pub fn is_active_blueprint(&self, blueprint_id: &StoreId) -> bool

Is the given blueprint id the active blueprint for any app id?

pub fn clear_active_blueprint(&mut self)

Clear the currently active blueprint

pub fn clear_active_blueprint_and_generate(&mut self)

Clear the currently active blueprint and enable the heuristics to generate a new one.

These keeps the default blueprint as is, so the user may reset to the default blueprint afterward.

pub fn clear_all_cloned_blueprints(&mut self)

Cloned blueprints are the ones the user has edited, i.e. NOT sent from the SDK.

pub fn purge_empty(&mut self)

Remove any empty EntityDbs from the hub

pub fn purge_fraction_of_ram(&mut self, fraction_to_purge: f32)

Call EntityDb::purge_fraction_of_ram on every recording

pub fn remove_recording_by_uri(&mut self, uri: &str)

Remove any recordings with a network source pointing at this uri.

pub fn gc_blueprints( &mut self, undo_state: &HashMap<StoreId, BlueprintUndoState>, )

pub fn begin_frame(&mut self, renderer_active_frame_idx: u64)

See re_viewer_context::Cache::begin_frame.

pub fn save_app_blueprints(&mut self) -> Result<()>

Persist any in-use blueprints to durable storage.

fn try_to_load_persisted_blueprint( &mut self, app_id: &ApplicationId, ) -> Result<()>

Try to load the persisted blueprint for the given ApplicationId. Note: If no blueprint exists at the expected path, the result is still considered Ok. It is only an Error if a blueprint exists but fails to load.

pub fn stats(&self) -> StoreHubStats

Populate a StoreHubStats based on the active app.