re_viewer_context
Module time_control

Structs

LastFrame ğŸ”’

TimeControl

Controls the global view and progress of the time.

TimeControlResponse

TimeState ğŸ”’

State per timeline.

TimeStateEntry ğŸ”’

TimeView

The time range we are currently zoomed in on.

Enums

ActiveTimeline ğŸ”’

Looping

PlayState

Functions

default_timeline ğŸ”’

Pick the timeline that should be the default, prioritizing user-defined ones.

max ğŸ”’

min ğŸ”’

range ğŸ”’

step_back_time ğŸ”’

step_back_time_looped ğŸ”’

step_fwd_time ğŸ”’

step_fwd_time_looped ğŸ”’