Trait re_viewer_context::DynSpaceViewClass

pub trait DynSpaceViewClass: Send + Sync {

    // Required methods
    fn identifier(&self) -> SpaceViewClassIdentifier;
    fn identifier_str() -> &'static str
       where Self: Sized;
    fn display_name(&self) -> &'static str;
    fn icon(&self) -> &'static Icon;
    fn help_text(&self, re_ui: &ReUi) -> WidgetText;
    fn on_register(
        &self,
        system_registry: &mut SpaceViewSystemRegistrator<'_>
    ) -> Result<(), SpaceViewClassRegistryError>;
    fn new_state(&self) -> Box<dyn SpaceViewState>;
    fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>;
    fn preferred_tile_aspect_ratio(
        &self,
        state: &dyn SpaceViewState
    ) -> Option<f32>;
    fn layout_priority(&self) -> SpaceViewClassLayoutPriority;
    fn visualizable_filter_context(
        &self,
        space_origin: &EntityPath,
        entity_db: &EntityDb
    ) -> Box<dyn VisualizableFilterContext>;
    fn auto_spawn_heuristic(
        &self,
        _ctx: &ViewerContext<'_>,
        space_origin: &EntityPath,
        ent_paths: &PerSystemEntities
    ) -> AutoSpawnHeuristic;
    fn on_frame_start(
        &self,
        ctx: &ViewerContext<'_>,
        state: &mut dyn SpaceViewState,
        ent_paths: &PerSystemEntities,
        entity_properties: &mut EntityPropertyMap
    );
    fn selection_ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut dyn SpaceViewState,
        space_origin: &EntityPath,
        space_view_id: SpaceViewId,
        root_entity_properties: &mut EntityProperties
    );
    fn ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut dyn SpaceViewState,
        root_entity_properties: &EntityProperties,
        query: &ViewQuery<'_>,
        system_output: SystemExecutionOutput
    );

}

Defines a class of space view without any concrete types making it suitable for storage and interfacing.

Implemented by crate::SpaceViewClass.

Each Space View in the viewer’s viewport has a single class assigned immutable at its creation time. The class defines all aspects of its behavior. It determines which entities are queried, how they are rendered, and how the user can interact with them.

TODO(andreas): Consider formulating a space view instance context object that is passed to all methods that operate on concrete space views as opposed to be about general information on the class.

Required Methods

fn identifier(&self) -> SpaceViewClassIdentifier

Identifier of this space view class.

Used for identification. Must be unique within a viewer session.

fn identifier_str() -> &'static strwhere Self: Sized,

Identifier string of this space view class.

This is similar to Self::identifier but it can be used without an actual instance of the class.

fn display_name(&self) -> &'static str

User-facing name of this space view class.

Used for UI display.

fn icon(&self) -> &'static Icon

Icon used to identify this space view class.

fn help_text(&self, re_ui: &ReUi) -> WidgetText

Help text describing how to interact with this space view in the ui.

fn on_register( &self, system_registry: &mut SpaceViewSystemRegistrator<'_> ) -> Result<(), SpaceViewClassRegistryError>

Called once upon registration of the class

This can be used to register all built-in crate::ViewContextSystem and crate::VisualizerSystem.

fn new_state(&self) -> Box<dyn SpaceViewState>

Called once for every new space view instance of this class.

The state is not persisted across viewer sessions, only shared frame-to-frame.

fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>

Optional archetype of the Space View’s blueprint properties.

Blueprint components that only apply to the space view itself, not to the entities it displays.

fn preferred_tile_aspect_ratio(&self, state: &dyn SpaceViewState) -> Option<f32>

Preferred aspect ratio for the ui tiles of this space view.

fn layout_priority(&self) -> SpaceViewClassLayoutPriority

Controls how likely this space view will get a large tile in the ui.

fn visualizable_filter_context( &self, space_origin: &EntityPath, entity_db: &EntityDb ) -> Box<dyn VisualizableFilterContext>

Create context object that is passed to all of this classes visualizers to determine whether they can be visualized

See crate::VisualizerSystem::filter_visualizable_entities.

fn auto_spawn_heuristic( &self, _ctx: &ViewerContext<'_>, space_origin: &EntityPath, ent_paths: &PerSystemEntities ) -> AutoSpawnHeuristic

Heuristic used to determine which space view is the best fit for a set of paths.

fn on_frame_start( &self, ctx: &ViewerContext<'_>, state: &mut dyn SpaceViewState, ent_paths: &PerSystemEntities, entity_properties: &mut EntityPropertyMap )

Executed for all active space views on frame start (before any ui is drawn), can be use for heuristic & state updates before populating the scene.

Is only allowed to access archetypes defined by Self::blueprint_archetype Passed entity properties are individual properties without propagated values.

fn selection_ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, state: &mut dyn SpaceViewState, space_origin: &EntityPath, space_view_id: SpaceViewId, root_entity_properties: &mut EntityProperties )

Ui shown when the user selects a space view of this class.

TODO(andreas): Should this be instead implemented via a registered data_ui of all blueprint relevant types?

fn ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, state: &mut dyn SpaceViewState, root_entity_properties: &EntityProperties, query: &ViewQuery<'_>, system_output: SystemExecutionOutput )

Draws the ui for this space view type and handles ui events.

The state passed in was previously created by Self::new_state and is kept frame-to-frame.