Trait re_viewer_context::SpaceViewClass

pub trait SpaceViewClass: Sized {
    type State: SpaceViewState + Default + 'static;

    // Required methods
    fn name(&self) -> SpaceViewClassName;
    fn icon(&self) -> &'static Icon;
    fn help_text(&self, re_ui: &ReUi) -> WidgetText;
    fn on_register(
        &self,
        system_registry: &mut SpaceViewSystemRegistry
    ) -> Result<(), SpaceViewClassRegistryError>;
    fn layout_priority(&self) -> SpaceViewClassLayoutPriority;
    fn selection_ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut Self::State,
        space_origin: &EntityPath,
        space_view_id: SpaceViewId
    );
    fn ui(
        &self,
        ctx: &mut ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut Self::State,
        view_ctx: &ViewContextCollection,
        parts: &ViewPartCollection,
        query: &ViewQuery<'_>,
        draw_data: Vec<QueueableDrawData>
    ) -> Result<(), SpaceViewSystemExecutionError>;

    // Provided methods
    fn preferred_tile_aspect_ratio(&self, _state: &Self::State) -> Option<f32> { ... }
    fn auto_spawn_heuristic(
        &self,
        _ctx: &ViewerContext<'_>,
        _space_origin: &EntityPath,
        ent_paths: &PerSystemEntities
    ) -> AutoSpawnHeuristic { ... }
    fn blueprint_archetype(&self) -> Option<Vec<ComponentName>> { ... }
    fn on_frame_start(
        &self,
        _ctx: &mut ViewerContext<'_>,
        _state: &Self::State,
        _ent_paths: &PerSystemEntities,
        _entity_properties: &mut EntityPropertyMap
    ) { ... }
}

Defines a class of space view.

Each Space View in the viewer’s viewport has a single class assigned immutable at its creation time. The class defines all aspects of its behavior. It determines which entities are queried, how they are rendered, and how the user can interact with them.

Required Associated Types

type State: SpaceViewState + Default + 'static

State of a space view.

Required Methods

fn name(&self) -> SpaceViewClassName

Name of this space view class.

Used for both ui display and identification. Must be unique within a viewer session.

fn icon(&self) -> &'static Icon

Icon used to identify this space view class.

fn help_text(&self, re_ui: &ReUi) -> WidgetText

Help text describing how to interact with this space view in the ui.

fn on_register( &self, system_registry: &mut SpaceViewSystemRegistry ) -> Result<(), SpaceViewClassRegistryError>

Called once upon registration of the class

This can be used to register all built-in crate::ViewContextSystem and crate::ViewPartSystem.

fn layout_priority(&self) -> SpaceViewClassLayoutPriority

Controls how likely this space view will get a large tile in the ui.

fn selection_ui( &self, ctx: &mut ViewerContext<'_>, ui: &mut Ui, state: &mut Self::State, space_origin: &EntityPath, space_view_id: SpaceViewId )

Ui shown when the user selects a space view of this class.

TODO(andreas): Should this be instead implemented via a registered data_ui of all blueprint relevant types?

fn ui( &self, ctx: &mut ViewerContext<'_>, ui: &mut Ui, state: &mut Self::State, view_ctx: &ViewContextCollection, parts: &ViewPartCollection, query: &ViewQuery<'_>, draw_data: Vec<QueueableDrawData> ) -> Result<(), SpaceViewSystemExecutionError>

Draws the ui for this space view class and handles ui events.

The passed state is kept frame-to-frame.

The passed systems (view_ctx and parts) are only valid for the duration of this frame and were already executed upon entering this method.

draw_data is all draw data gathered by executing the view part systems. TODO(wumpf): Right now the ui methods control when and how to create re_renderer::ViewBuilders. In the future, we likely want to move view builder handling to re_viewport with minimal configuration options exposed via crate::SpaceViewClass.

Provided Methods

fn preferred_tile_aspect_ratio(&self, _state: &Self::State) -> Option<f32>

Preferred aspect ratio for the ui tiles of this space view.

fn auto_spawn_heuristic( &self, _ctx: &ViewerContext<'_>, _space_origin: &EntityPath, ent_paths: &PerSystemEntities ) -> AutoSpawnHeuristic

Heuristic used to determine which space view is the best fit for a set of paths.

For each path in ent_paths, at least one of the registered crate::ViewPartSystem for this class returned true when calling crate::ViewPartSystem::heuristic_filter.

fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>

Optional archetype of the Space View’s blueprint properties.

Blueprint components that only apply to the space view itself, not to the entities it displays.

fn on_frame_start( &self, _ctx: &mut ViewerContext<'_>, _state: &Self::State, _ent_paths: &PerSystemEntities, _entity_properties: &mut EntityPropertyMap )

Executed for all active space views on frame start (before any ui is drawn), can be use for heuristic & state updates before populating the scene.

Is only allowed to access archetypes defined by Self::blueprint_archetype Passed entity properties are individual properties without propagated values.