Trait re_viewer_context::SpaceViewState

pub trait SpaceViewState: Any {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;
}

Unserialized frame to frame state of a space view.

For any state that should be persisted, use the Blueprint! This state is used for transient state, such as animation or uncommitted ui state like dragging a camera. (on mouse release, the camera would be committed to the blueprint).

Required Methods

fn as_any(&self) -> &dyn Any

Converts itself to a reference of std::any::Any, which enables downcasting to concrete types.

fn as_any_mut(&mut self) -> &mut dyn Any

Converts itself to a reference of std::any::Any, which enables downcasting to concrete types.