Trait re_viewer_context::ViewPartSystem

pub trait ViewPartSystem {
    // Required methods
    fn required_components(&self) -> ComponentNameSet;
    fn execute(
        &mut self,
        ctx: &mut ViewerContext<'_>,
        query: &ViewQuery<'_>,
        view_ctx: &ViewContextCollection
    ) -> Result<Vec<QueueableDrawData>, SpaceViewSystemExecutionError>;
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn indicator_components(&self) -> ComponentNameSet { ... }
    fn heuristic_filter(
        &self,
        _store: &DataStore,
        _ent_path: &EntityPath,
        _query: &LatestAtQuery,
        entity_components: &ComponentNameSet
    ) -> bool { ... }
    fn data(&self) -> Option<&dyn Any> { ... }
}

Element of a scene derived from a single archetype query.

Is populated after scene contexts and has access to them.

Required Methods

fn required_components(&self) -> ComponentNameSet

Returns the minimal set of components that the system requires in order to be instantiated.

This does not include indicator components.

fn execute( &mut self, ctx: &mut ViewerContext<'_>, query: &ViewQuery<'_>, view_ctx: &ViewContextCollection ) -> Result<Vec<QueueableDrawData>, SpaceViewSystemExecutionError>

Queries the data store and performs data conversions to make it ready for display.

Musn’t query any data outside of the archetype.

TODO(andreas): don’t pass in ViewerContext if we want to restrict the queries here. If we want to make this restriction, then the trait-contract should be that something external to the ViewPartSystemImpl does the query and then passes an ArchetypeQueryResult into populate.

fn as_any(&self) -> &dyn Any

Provided Methods

fn indicator_components(&self) -> ComponentNameSet

These are not required, but if any of these are found, it is a strong indication that this system should be active (if also the required_components are found).

fn heuristic_filter( &self, _store: &DataStore, _ent_path: &EntityPath, _query: &LatestAtQuery, entity_components: &ComponentNameSet ) -> bool

Implements a filter to heuristically determine whether or not to instantiate the system.

If and when the system can be instantiated (i.e. because there is at least one entity that satisfies the minimal set of required components), this method applies an arbitrary filter to determine whether or not the system should be instantiated by default.

The passed-in set of entity_components corresponds to all the different components that have ever been logged on the entity path.

By default, this returns true if eiher Self::indicator_components is empty or entity_components contains at least one of these indicator components.

Override this method only if a more detailed condition is required to inform heuristics whether or not the given entity is relevant for this system.

fn data(&self) -> Option<&dyn Any>

Optionally retrieves a data store reference from the scene element.

This is useful for retrieving data that is common to several parts of a crate::SpaceViewClass. For example, if most scene parts produce ui elements, a concrete crate::SpaceViewClass can pick those up in its crate::SpaceViewClass::ui method by iterating over all scene parts.