re_viewer_context
Trait VisualizerSystem

pub trait VisualizerSystem:
    Send
    + Sync
    + 'static {
    // Required methods
    fn visualizer_query_info(&self) -> VisualizerQueryInfo;
    fn execute(
        &mut self,
        ctx: &ViewContext<'_>,
        query: &ViewQuery<'_>,
        context_systems: &ViewContextCollection,
    ) -> Result<Vec<QueueableDrawData>, ViewSystemExecutionError>;
    fn as_any(&self) -> &dyn Any;
    fn fallback_provider(&self) -> &dyn ComponentFallbackProvider;

    // Provided methods
    fn filter_visualizable_entities(
        &self,
        entities: MaybeVisualizableEntities,
        _context: &dyn VisualizableFilterContext,
    ) -> VisualizableEntities { ... }
    fn data_based_visualizability_filter(
        &self,
    ) -> Option<Box<dyn DataBasedVisualizabilityFilter>> { ... }
    fn data(&self) -> Option<&dyn Any> { ... }
}

Element of a scene derived from a single archetype query.

Is populated after scene contexts and has access to them.

All visualizers are expected to be able to provide a fallback value for any component they’re using via the ComponentFallbackProvider trait.

Required Methods

fn visualizer_query_info(&self) -> VisualizerQueryInfo

Information about which components are queried by the visualizer.

fn execute( &mut self, ctx: &ViewContext<'_>, query: &ViewQuery<'_>, context_systems: &ViewContextCollection, ) -> Result<Vec<QueueableDrawData>, ViewSystemExecutionError>

Queries the chunk store and performs data conversions to make it ready for display.

Mustn’t query any data outside of the archetype.

fn as_any(&self) -> &dyn Any

fn fallback_provider(&self) -> &dyn ComponentFallbackProvider

Returns the fallback provider for this visualizer.

Visualizers should use this to report the fallback values they use when there is no data. The Rerun viewer will display these fallback values to the user to convey what the visualizer is doing.

Provided Methods

fn filter_visualizable_entities( &self, entities: MaybeVisualizableEntities, _context: &dyn VisualizableFilterContext, ) -> VisualizableEntities

Filters a set of “maybe visualizable” entities (entities that have all required components and fulfill other view independent criteria), into to a set of “visualizable” entities.

The context passed in here is generated by crate::ViewClass::visualizable_filter_context.

fn data_based_visualizability_filter( &self, ) -> Option<Box<dyn DataBasedVisualizabilityFilter>>

Additional filter for visualizability based on component data.

If none is specified, “maybe visualizable” is solely determined by required components. (for final visualizability, the view instance dependent filter is applied, see crate::VisualizerSystem::filter_visualizable_entities)

fn data(&self) -> Option<&dyn Any>

Optionally retrieves a chunk store reference from the scene element.

This is useful for retrieving data that is common to several visualizers of a crate::ViewClass. For example, if most visualizers produce ui elements, a concrete crate::ViewClass can pick those up in its crate::ViewClass::ui method by iterating over all visualizers.