Struct re_viewer_context::typed_entity_collections::IndicatedEntities

pub struct IndicatedEntities(pub IntSet<EntityPath>);

List of entities that match the indicator components of a visualizer.

In order to be a match the entity must have at some point in time on any timeline had any of the indicator components specified by the respective visualizer system.

Tuple Fields
0: IntSet<EntityPath>