re_viewer_context::view::named_system
Trait IdentifiedViewSystem

pub trait IdentifiedViewSystem {
    // Required method
    fn identifier() -> ViewSystemIdentifier;
}

Trait for naming/identifying crate::VisualizerSystems & crate::ViewContextSystems.

Required to be implemented for registration.

Required Methods

fn identifier() -> ViewSystemIdentifier

Unique name for a system within a given crate::ViewClass.

Note that this is not unique across the entire application.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.