Trait re_viewer_context::view::view_class::ViewStateExt

pub trait ViewStateExt: ViewState {
    // Provided methods
    fn downcast_ref<T: ViewState>(&self) -> Result<&T, ViewSystemExecutionError> { ... }
    fn downcast_mut<T: ViewState>(
        &mut self,
    ) -> Result<&mut T, ViewSystemExecutionError> { ... }
}

Provided Methods

fn downcast_ref<T: ViewState>(&self) -> Result<&T, ViewSystemExecutionError>

Downcasts this state to a reference of a concrete type.

fn downcast_mut<T: ViewState>( &mut self, ) -> Result<&mut T, ViewSystemExecutionError>

Downcasts this state to a mutable reference of a concrete type.

Object Safety

This trait is not object safe.