re_viewer_context
Module component_fallbacks

Enums

ComponentFallbackError

Error type for a fallback request.

ComponentFallbackProviderResult

Result for a fallback request to a provider.

Traits

ComponentFallbackProvider

Provides fallback values for components, implemented typically by crate::ViewClass and crate::VisualizerSystem.

TypedComponentFallbackProvider

Provides a fallback value for a given component with known type.