Module re_viewer_context::time_control

Structs

TimeControl

Controls the global view and progress of the time.

TimeState ğŸ”’

State per timeline.

TimeView

The time range we are currently zoomed in on.

Enums

Looping

PlayState

Functions

default_timeline ğŸ”’

Pick the timeline that should be the default, prioritizing user-defined ones.

max ğŸ”’

min ğŸ”’

range ğŸ”’

step_back_time ğŸ”’

step_back_time_looped ğŸ”’

step_fwd_time ğŸ”’

step_fwd_time_looped ğŸ”’