Trait re_viewer_context::ViewClass

pub trait ViewClass: Send + Sync {

    // Required methods
    fn identifier() -> ViewClassIdentifier
       where Self: Sized;
    fn display_name(&self) -> &'static str;
    fn help(&self, egui_ctx: &Context) -> Help<'_>;
    fn on_register(
        &self,
        system_registry: &mut ViewSystemRegistrator<'_>,
    ) -> Result<(), ViewClassRegistryError>;
    fn new_state(&self) -> Box<dyn ViewState>;
    fn layout_priority(&self) -> ViewClassLayoutPriority;
    fn spawn_heuristics(&self, ctx: &ViewerContext<'_>) -> ViewSpawnHeuristics;
    fn ui(
        &self,
        ctx: &ViewerContext<'_>,
        ui: &mut Ui,
        state: &mut dyn ViewState,
        query: &ViewQuery<'_>,
        system_output: SystemExecutionOutput,
    ) -> Result<(), ViewSystemExecutionError>;

    // Provided methods
    fn icon(&self) -> &'static Icon { ... }
    fn blueprint_archetype(&self) -> Option<Vec<ComponentName>> { ... }
    fn preferred_tile_aspect_ratio(&self, _state: &dyn ViewState) -> Option<f32> { ... }
    fn supports_visible_time_range(&self) -> bool { ... }
    fn default_query_range(&self, _state: &dyn ViewState) -> QueryRange { ... }
    fn recommended_root_for_entities(
        &self,
        _entities: &IntSet<EntityPath>,
        _entity_db: &EntityDb,
    ) -> Option<EntityPath> { ... }
    fn visualizable_filter_context(
        &self,
        _space_origin: &EntityPath,
        _entity_db: &EntityDb,
    ) -> Box<dyn VisualizableFilterContext> { ... }
    fn choose_default_visualizers(
        &self,
        entity_path: &EntityPath,
        _maybe_visualizable_entities_per_visualizer: &PerVisualizer<MaybeVisualizableEntities>,
        visualizable_entities_per_visualizer: &PerVisualizer<VisualizableEntities>,
        indicated_entities_per_visualizer: &PerVisualizer<IndicatedEntities>,
    ) -> SmallVisualizerSet { ... }
    fn selection_ui(
        &self,
        _ctx: &ViewerContext<'_>,
        _ui: &mut Ui,
        _state: &mut dyn ViewState,
        _space_origin: &EntityPath,
        _view_id: ViewId,
    ) -> Result<(), ViewSystemExecutionError> { ... }
    fn extra_title_bar_ui(
        &self,
        _ctx: &ViewerContext<'_>,
        _ui: &mut Ui,
        _state: &mut dyn ViewState,
        _space_origin: &EntityPath,
        _view_id: ViewId,
    ) -> Result<(), ViewSystemExecutionError> { ... }

}

Defines a class of view without any concrete types making it suitable for storage and interfacing.

Each View in the viewer’s viewport has a single class assigned immutable at its creation time. The class defines all aspects of its behavior. It determines which entities are queried, how they are rendered, and how the user can interact with them.

Required Methods

fn identifier() -> ViewClassIdentifier
where Self: Sized,

Identifier string of this view class.

By convention we use PascalCase.

fn display_name(&self) -> &'static str

User-facing name of this view class.

Used for UI display.

fn help(&self, egui_ctx: &Context) -> Help<'_>

fn on_register( &self, system_registry: &mut ViewSystemRegistrator<'_>, ) -> Result<(), ViewClassRegistryError>

Called once upon registration of the class

This can be used to register all built-in crate::ViewContextSystem and crate::VisualizerSystem.

fn new_state(&self) -> Box<dyn ViewState>

Called once for every new view instance of this class.

The state is not persisted across viewer sessions, only shared frame-to-frame.

fn layout_priority(&self) -> ViewClassLayoutPriority

Controls how likely this view will get a large tile in the ui.

fn spawn_heuristics(&self, ctx: &ViewerContext<'_>) -> ViewSpawnHeuristics

Determines which views should be spawned by default for this class.

fn ui( &self, ctx: &ViewerContext<'_>, ui: &mut Ui, state: &mut dyn ViewState, query: &ViewQuery<'_>, system_output: SystemExecutionOutput, ) -> Result<(), ViewSystemExecutionError>

Draws the ui for this view class and handles ui events.

The passed state is kept frame-to-frame.

TODO(wumpf): Right now the ui methods control when and how to create re_renderer::ViewBuilders. In the future, we likely want to move view builder handling to re_viewport with minimal configuration options exposed via crate::ViewClass.

Provided Methods

fn icon(&self) -> &'static Icon

Icon used to identify this view class.

fn blueprint_archetype(&self) -> Option<Vec<ComponentName>>

Optional archetype of the View’s blueprint properties.

Blueprint components that only apply to the view itself, not to the entities it displays.

fn preferred_tile_aspect_ratio(&self, _state: &dyn ViewState) -> Option<f32>

Preferred aspect ratio for the ui tiles of this view.

fn supports_visible_time_range(&self) -> bool

Controls whether the visible time range UI should be displayed for this view.

fn default_query_range(&self, _state: &dyn ViewState) -> QueryRange

Default query range for this view.

fn recommended_root_for_entities( &self, _entities: &IntSet<EntityPath>, _entity_db: &EntityDb, ) -> Option<EntityPath>

Determines a suitable origin given the provided set of entities.

This function only considers the transform topology, disregarding the actual visualizability of the entities (for this, use Self::visualizable_filter_context).

fn visualizable_filter_context( &self, _space_origin: &EntityPath, _entity_db: &EntityDb, ) -> Box<dyn VisualizableFilterContext>

Create context object that is passed to all of this classes visualizers to determine whether they can be visualized

See crate::VisualizerSystem::filter_visualizable_entities.

fn choose_default_visualizers( &self, entity_path: &EntityPath, _maybe_visualizable_entities_per_visualizer: &PerVisualizer<MaybeVisualizableEntities>, visualizable_entities_per_visualizer: &PerVisualizer<VisualizableEntities>, indicated_entities_per_visualizer: &PerVisualizer<IndicatedEntities>, ) -> SmallVisualizerSet

Choose the default visualizers to enable for this entity.

Helpful for customizing fallback behavior for types that are insufficient to determine indicated on their own.

Will only be called for entities where the selected visualizers have not been overridden by the blueprint.

This interface provides a default implementation which will return all visualizers which are both visualizable and indicated for the given entity.

fn selection_ui( &self, _ctx: &ViewerContext<'_>, _ui: &mut Ui, _state: &mut dyn ViewState, _space_origin: &EntityPath, _view_id: ViewId, ) -> Result<(), ViewSystemExecutionError>

Ui shown when the user selects a view of this class.

fn extra_title_bar_ui( &self, _ctx: &ViewerContext<'_>, _ui: &mut Ui, _state: &mut dyn ViewState, _space_origin: &EntityPath, _view_id: ViewId, ) -> Result<(), ViewSystemExecutionError>

Additional UI displayed in the tab title bar, between the “maximize” and “help” buttons.

Note: this is a right-to-left layout.