re_viewer_context::view::view_class
Trait ViewStateExt

pub trait ViewStateExt: ViewState {
    // Provided methods
    fn downcast_ref<T: ViewState>(&self) -> Result<&T, ViewSystemExecutionError> { ... }
    fn downcast_mut<T: ViewState>(
        &mut self,
    ) -> Result<&mut T, ViewSystemExecutionError> { ... }
}

Provided Methods

fn downcast_ref<T: ViewState>(&self) -> Result<&T, ViewSystemExecutionError>

Downcasts this state to a reference of a concrete type.

fn downcast_mut<T: ViewState>( &mut self, ) -> Result<&mut T, ViewSystemExecutionError>

Downcasts this state to a mutable reference of a concrete type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.